fix(security): address Copilot review findings from PR #292
C1: Split csrf.ts into server-only (crypto/NextRequest) and
    csrf-client.ts (Web Crypto API) to prevent client bundle breakage
C2: Add @naap/crypto to transpilePackages in next.config.js
C3: Guard verifyPassword timingSafeEqual with try/catch and buffer
    length check to handle malformed hex in stored hashes
C4: base-svc createSession now sets versionAtCreation from user
C5: base-svc validateSession, validateSessionWithExpiry, and
    refreshSession now enforce sessionVersion vs versionAtCreation
C6: hmacToken requires SESSION_TOKEN_PEPPER in production (throws),
    allows SHA-256 fallback only in dev/test — all three locations
C7: verify-token-hashing.sql runbook updated to match dual-write plan
    (token column stays non-nullable during rollout)
C8: neon-roles.sql uses GRANT CREATE ON DATABASE instead of CREATEDB
    for least-privilege migrator role
