+++ apps/web-next/next.config.js
@@ -27,6 +27,7 @@ const nextConfig = {
     '@naap/config',
     '@naap/plugin-sdk',
     '@naap/cache',
+    '@naap/crypto',
   ],
 
   // Image optimization
