+++ apps/web-next/src/lib/api/auth.ts
@@ -23,6 +23,9 @@ function sanitizeForLog(value: unknown): string {
 function hmacToken(plaintext: string): string {
   const pepper = process.env.SESSION_TOKEN_PEPPER;
   if (!pepper) {
+    if (process.env.NODE_ENV === 'production') {
+      throw new Error('SESSION_TOKEN_PEPPER is required in production');
+    }
     return crypto.createHash('sha256').update(plaintext).digest('hex');
   }
   return crypto.createHmac('sha256', pepper).update(plaintext).digest('hex');
@@ -69,14 +72,16 @@ async function verifyPassword(password: string, storedHash: string): Promise<{ v
 
   if (!salt || !hash) return { valid: false, needsRehash: false };
   const verifyHash = crypto.pbkdf2Sync(password, salt, iterations, 64, digest).toString('hex');
-  if (hash.length !== verifyHash.length) return { valid: false, needsRehash: false };
 
-  const valid = crypto.timingSafeEqual(
-    Buffer.from(hash, 'hex'),
-    Buffer.from(verifyHash, 'hex')
-  );
-
-  return { valid, needsRehash: valid && needsRehash };
+  try {
+    const storedBuf = Buffer.from(hash, 'hex');
+    const verifyBuf = Buffer.from(verifyHash, 'hex');
+    if (storedBuf.length !== verifyBuf.length) return { valid: false, needsRehash: false };
+    const valid = crypto.timingSafeEqual(storedBuf, verifyBuf);
+    return { valid, needsRehash: valid && needsRehash };
+  } catch {
+    return { valid: false, needsRehash: false };
+  }
 }
 
 function generateToken(): string {
