+++ apps/web-next/src/lib/api/index.ts
@@ -5,3 +5,4 @@
 export * from './response';
 export * from './auth';
 export * from './csrf';
+export * from './csrf-client';
